修改撤销买单定时任务逻辑

+++ src/main/java/com/slabs/exchange/service/back/IProjectService.java
@@ -20,7 +20,7 @@ public interface IProjectService {
 
     ResponseBean getForeProjectList(PageParamDto pageParamDto);
 
-    ResponseBean getForeProjectDetail(Long projectId);
+    ResponseBean getForeProjectDetail(Integer projectId);
 
     ResponseBean buy(BuyDto buyDto);
 }
